Rename `cugraph-nx` to `nx-cugraph`
