+++ ci/build_python.sh
@@ -26,7 +26,7 @@ rapids-mamba-retry mambabuild \
   --channel "${RAPIDS_CONDA_BLD_OUTPUT_DIR}" \
   conda/recipes/cugraph
 
-# NOTE: nothing in cugraph-nx is CUDA-specific, but it is built on each CUDA
+# NOTE: nothing in nx-cugraph is CUDA-specific, but it is built on each CUDA
 # platform to ensure it is included in each set of artifacts, since test
 # scripts only install from one set of artifacts based on the CUDA version used
 # for the test run.
@@ -34,7 +34,7 @@ rapids-mamba-retry mambabuild \
   --no-test \
   --channel "${CPP_CHANNEL}" \
   --channel "${RAPIDS_CONDA_BLD_OUTPUT_DIR}" \
-  conda/recipes/cugraph-nx
+  conda/recipes/nx-cugraph
 
 # NOTE: nothing in the cugraph-service packages are CUDA-specific, but they are
 # built on each CUDA platform to ensure they are included in each set of
