+++ conda/recipes/nx-cugraph/build.sh
@@ -4,4 +4,4 @@
 
 # This assumes the script is executed from the root of the repo directory
 
-./build.sh cugraph-nx
+./build.sh nx-cugraph
