+++ python/cugraph/cugraph/tests/link_analysis/test_pagerank.py
@@ -65,7 +65,7 @@ def cugraph_call(G, max_iter, tol, alpha, personalization, nstart, pre_vtx_o_wgt
 
 
 # need a different function since the Nx version returns a dictionary
-def cugraph_nx_call(G, max_iter, tol, alpha, personalization, nstart):
+def nx_cugraph_call(G, max_iter, tol, alpha, personalization, nstart):
     # cugraph Pagerank Call
     t1 = time.time()
     pr = cugraph.pagerank(
@@ -238,7 +238,7 @@ def test_pagerank_nx(graph_file, max_iter, tol, alpha, personalization_perc, has
     cu_prsn = cudify(networkx_prsn)
 
     # cuGraph PageRank with Nx Graph
-    cugraph_pr = cugraph_nx_call(Gnx, max_iter, tol, alpha, cu_prsn, cu_nstart)
+    cugraph_pr = nx_cugraph_call(Gnx, max_iter, tol, alpha, cu_prsn, cu_nstart)
 
     # Calculating mismatch
     networkx_pr = sorted(networkx_pr.items(), key=lambda x: x[0])
