+++ python/nx-cugraph/.flake8
@@ -9,5 +9,5 @@ extend-ignore =
     SIM401,
 # E203 whitespace before ':' (to be compatible with black)
 per-file-ignores =
-    cugraph_nx/tests/*.py:T201,
+    nx_cugraph/tests/*.py:T201,
     __init__.py:F401,F403,
