+++ python/nx-cugraph/README.md
@@ -1,24 +1,24 @@
-# cugraph-nx
+# nx-cugraph
 
 ## Description
-[RAPIDS](https://rapids.ai) cugraph-nx is a [backend to NetworkX](https://networkx.org/documentation/stable/reference/classes/index.html#backends)
+[RAPIDS](https://rapids.ai) nx-cugraph is a [backend to NetworkX](https://networkx.org/documentation/stable/reference/classes/index.html#backends)
 with minimal dependencies (`networkx`, `cupy`, and `pylibcugraph`) to run graph algorithms on the GPU.
 
 ### Contribute
 
 Follow instructions for [contributing to cugraph](https://github.com/rapidsai/cugraph/blob/branch-23.10/readme_pages/CONTRIBUTING.md)
-and [building from source](https://docs.rapids.ai/api/cugraph/stable/installation/source_build/), then build cugraph-nx in develop (i.e., editable) mode:
+and [building from source](https://docs.rapids.ai/api/cugraph/stable/installation/source_build/), then build nx-cugraph in develop (i.e., editable) mode:
 ```
-$ ./build.sh cugraph-nx --pydevelop
+$ ./build.sh nx-cugraph --pydevelop
 ```
 
 ### Run tests
 
-Run cugraph-nx tests from `cugraph/python/cugraph-nx` directory:
+Run nx-cugraph tests from `cugraph/python/nx-cugraph` directory:
 ```
 $ pytest
 ```
-Run cugraph-nx benchmarks:
+Run nx-cugraph benchmarks:
 ```
 $ pytest --bench
 ```
