+++ python/nx-cugraph/nx_cugraph/algorithms/community/louvain.py
@@ -14,8 +14,8 @@ import sys
 
 import pylibcugraph as plc
 
-from cugraph_nx.convert import _to_undirected_graph
-from cugraph_nx.utils import _groupby, networkx_algorithm, not_implemented_for
+from nx_cugraph.convert import _to_undirected_graph
+from nx_cugraph.utils import _groupby, networkx_algorithm, not_implemented_for
 
 __all__ = ["louvain_communities"]
