+++ python/nx-cugraph/nx_cugraph/utils/misc.py
