+++ python/nx-cugraph/pyproject.toml
@@ -9,7 +9,7 @@ requires = [
 build-backend = "setuptools.build_meta"
 
 [project]
-name = "cugraph-nx"
+name = "nx-cugraph"
 version = "23.10.00"
 description = "cugraph backend for NetworkX"
 readme = { file = "README.md", content-type = "text/markdown" }
@@ -52,15 +52,15 @@ Homepage = "https://github.com/rapidsai/cugraph"
 Documentation = "https://docs.rapids.ai/api/cugraph/stable/"
 
 [project.entry-points."networkx.plugins"]
-cugraph = "cugraph_nx.interface:BackendInterface"
+cugraph = "nx_cugraph.interface:BackendInterface"
 
 [tool.setuptools]
 license-files = ["LICENSE"]
 
 [tool.setuptools.packages.find]
 include = [
-    "cugraph_nx*",
-    "cugraph_nx.*",
+    "nx_cugraph*",
+    "nx_cugraph.*",
 ]
 
 [tool.black]
@@ -73,12 +73,12 @@ profile = "black"
 skip_gitignore = true
 float_to_top = true
 default_section = "THIRDPARTY"
-known_first_party = "cugraph_nx"
+known_first_party = "nx_cugraph"
 line_length = 88
 
 [tool.pytest.ini_options]
 minversion = "6.0"
-testpaths = "cugraph_nx/tests"
+testpaths = "nx_cugraph/tests"
 xfail_strict = true
 markers = [
     "slow: Skipped unless --runslow passed",
@@ -109,7 +109,7 @@ addopts = [
 
 [tool.coverage.run]
 branch = true
-source = ["cugraph_nx"]
+source = ["nx_cugraph"]
 omit = []
 
 [tool.coverage.report]
@@ -202,7 +202,7 @@ ignore = [
 [tool.ruff.per-file-ignores]
 "__init__.py" = ["F401"]  # Allow unused imports (w/o defining `__all__`)
 # Allow assert, print, RNG, and no docstring
-"cugraph_nx/**/tests/*py" = ["S101", "S311", "T201", "D103", "D100"]
+"nx_cugraph/**/tests/*py" = ["S101", "S311", "T201", "D103", "D100"]
 
 [tool.ruff.flake8-annotations]
 mypy-init-return = true
