+++ python/nx-cugraph/run_nx_tests.sh
@@ -4,21 +4,21 @@
 #
 # NETWORKX_GRAPH_CONVERT=cugraph
 #   Used by networkx versions 3.0 and 3.1
-#   Must be set to "cugraph" to test the cugraph-nx backend.
+#   Must be set to "cugraph" to test the nx-cugraph backend.
 #
 # NETWORKX_TEST_BACKEND=cugraph
 #   Replaces NETWORKX_GRAPH_CONVERT for networkx versions >=3.2
-#   Must be set to "cugraph" to test the cugraph-nx backend.
+#   Must be set to "cugraph" to test the nx-cugraph backend.
 #
 # NETWORKX_FALLBACK_TO_NX=True (optional)
 #   Used by networkx versions >=3.2.  With this set, input graphs will not be
-#   converted to cugraph-nx and the networkx algorithm will be called for
+#   converted to nx-cugraph and the networkx algorithm will be called for
 #   algorithms that we don't implement or if we raise NotImplementedError.
 #   This is sometimes helpful to get increased testing and coverage, but
 #   testing takes longer.  Without it, tests will xfail when encountering a
 #   function that we don't implement.
 #
-# Coverage of `cugraph_nx.algorithms` is reported and is a good sanity check
+# Coverage of `nx_cugraph.algorithms` is reported and is a good sanity check
 # that algorithms run.
 
 # Warning: cugraph has a .coveragerc file in the <repo root>/python directory,
@@ -30,7 +30,7 @@ NETWORKX_TEST_BACKEND=cugraph \
 NETWORKX_FALLBACK_TO_NX=True \
     pytest \
     --pyargs networkx \
-    --cov=cugraph_nx.algorithms \
+    --cov=nx_cugraph.algorithms \
     --cov-report term-missing \
     --no-cov-on-fail \
     "$@"
